Move create_full_handle_for_index into handle_creation_tests.rs
