+++ pallets/handles/src/tests/handle_creation_tests.rs
@@ -1,16 +1,47 @@
 use crate::{tests::mock::*, Error, Event};
 use codec::Decode;
-use common_primitives::msa::MessageSourceId;
+use common_primitives::{handles::SequenceIndex, msa::MessageSourceId};
 use frame_support::{assert_noop, assert_ok};
+use handles_utils::converter::convert_to_canonical;
 use sp_core::{sr25519, Encode, Pair};
 use sp_std::collections::btree_set::BTreeSet;
+
+/// Creates a full display handle by combining a base handle string with a suffix generated
+/// from an index into the suffix sequence.
+///
+/// # Arguments
+///
+/// * `base_handle_str` - The base handle string.
+/// * `suffix_sequence_index` - The index into the suffix sequence.
+///
+/// # Returns
+///
+/// * `DisplayHandle` - The full display handle.
+///
+fn create_full_handle_for_index(
+	base_handle_str: &str,
+	suffix_sequence_index: SequenceIndex,
+) -> Vec<u8> {
+	// Convert base handle into a canonical base
+	let canonical_handle_str = convert_to_canonical(&base_handle_str);
+
+	// Generate suffix from index into the suffix sequence
+	let suffix = Handles::generate_suffix_for_canonical_handle(
+		&canonical_handle_str,
+		suffix_sequence_index as usize,
+	)
+	.unwrap_or_default();
+
+	let display_handle = Handles::build_full_display_handle(base_handle_str, suffix).unwrap();
+	display_handle.into_inner()
+}
 
 #[test]
 fn test_full_handle_creation() {
 	new_test_ext().execute_with(|| {
 		// Min is 10, Max is 99 inclusive
 		for sequence_index in 0..89 {
-			let display_handle = Handles::create_full_handle_for_index("test", sequence_index);
+			let display_handle = create_full_handle_for_index("test", sequence_index);
 			assert_ok!(core::str::from_utf8(&display_handle));
 		}
 	})
@@ -32,7 +63,7 @@ fn claim_handle_happy_path() {
 
 		// Confirm that HandleClaimed event was deposited
 		let msa_id = MessageSourceId::decode(&mut &alice.public().encode()[..]).unwrap();
-		let handle = Handles::create_full_handle_for_index("test1", 0);
+		let handle = create_full_handle_for_index("test1", 0);
 		System::assert_last_event(Event::HandleClaimed { msa_id, handle }.into());
 	});
 }
